requirements.txt
psycopg-binary==3.2.3
psycopg2-binary==2.9.9
